move config to cmd dir
